Separate API routes and 404 page
