+++ modules/webServer.js
@@ -1,4 +1,5 @@
 const express = require("express");
+const { getAll } = require("./database");
 
 // Create the express server and attach socket.io to it
 function createServer() {
@@ -9,5 +10,13 @@ function createServer() {
     return { app, io, http };
 }
 
+async function getIpAccess(type) {
+	const ipList = await getAll(`SELECT id, ip FROM ip_${type}`)
+	return ipList.reduce((ips, ip) => {
+		ips[ip.id] = ip
+		return ips
+	}, {})
+}
+
 const { app, io, http } = createServer();
-module.exports = { app, io, http };
+module.exports = { app, io, http, getIpAccess };
